+++ database/migrations/2021_05_05_201643_create_availables_table.php
@@ -21,8 +21,8 @@ class CreateAvailablesTable extends Migration
             $table->date('e_date');
             $table->string('booking_name');
             $table->boolean('status');
-            $table->double('bill');
-            $table->double('paid');
+            $table->bigInteger('bill');
+            $table->bigInteger('paid');
             $table->timestamps();
         });
     }
